LinkedIn-Automation-Frontend-main/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
numpy
pandas
Pillow
psycopg2
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
django-rest-framework==0.1.0
plotly==5.20.0
django-simple-search==1.0.2
xlsxwriter==3.2.0
